+++ src/rules/rebase_commit.rs
@@ -13,32 +13,40 @@ impl RebaseCommit {
 
     pub fn validate(&self, commit: &Commit) -> Option<Vec<Issue>> {
         let subject = &commit.subject;
-        if subject.starts_with("fixup! ") {
-            let context = Context::subject_error(
-                subject.to_string(),
-                Range { start: 0, end: 6 },
-                "Rebase fixup commits before pushing or merging".to_string(),
-            );
-            return Some(vec![Issue::error(
-                Rule::RebaseCommit,
-                "A fixup commit was found".to_string(),
-                Position::Subject { line: 1, column: 1 },
-                vec![context],
-            )]);
-        } else if subject.starts_with("squash! ") {
-            let context = Context::subject_error(
-                subject.to_string(),
-                Range { start: 0, end: 7 },
-                "Rebase squash commits before pushing or merging".to_string(),
-            );
-            return Some(vec![Issue::error(
-                Rule::RebaseCommit,
-                "A squash commit was found".to_string(),
-                Position::Subject { line: 1, column: 1 },
-                vec![context],
-            )]);
+        let fixup_check = validate_prefix("fixup", subject);
+        if fixup_check.is_some() {
+            return fixup_check;
+        }
+        let squash_check = validate_prefix("squash", subject);
+        if squash_check.is_some() {
+            return squash_check;
+        }
+        let amend_check = validate_prefix("amend", subject);
+        if amend_check.is_some() {
+            return amend_check;
         }
 
+        None
+    }
+}
+
+fn validate_prefix(prefix: &str, subject: &str) -> Option<Vec<Issue>> {
+    if subject.starts_with(&format!("{}! ", prefix)) {
+        let context = Context::subject_error(
+            subject.to_string(),
+            Range {
+                start: 0,
+                end: prefix.len() + 1,
+            },
+            format!("Rebase {} commits before pushing or merging", prefix),
+        );
+        Some(vec![Issue::error(
+            Rule::RebaseCommit,
+            format!("A {} commit was found", prefix),
+            Position::Subject { line: 1, column: 1 },
+            vec![context],
+        )])
+    } else {
         None
     }
 }
@@ -82,6 +90,18 @@ mod tests {
             &issue,
             "1 | squash! I need a rebase\n\
                | ^^^^^^^ Rebase squash commits before pushing or merging",
+        );
+    }
+
+    #[test]
+    fn with_amend_commit() {
+        let issue = first_issue(validate(&commit("amend! I need a rebase", "")));
+        assert_eq!(issue.message, "A amend commit was found");
+        assert_eq!(issue.position, subject_position(1));
+        assert_contains_issue_output(
+            &issue,
+            "1 | amend! I need a rebase\n\
+               | ^^^^^^ Rebase amend commits before pushing or merging",
         );
     }
 }